+++ src/main/java/com/example/demo/entities/Direction.java
@@ -30,7 +30,7 @@ public class Direction {
     private String treepath;
 
     @Column(name = "parentorganizationId")
-    private Integer parentorganizationId;
+    private Long parentorganizationId;
 
     @Size(max = 255)
     @NotNull
@@ -39,7 +39,7 @@ public class Direction {
 
     public Direction() {
     }
-    public Direction(Long id, Long organizationId, Long level, String type, String treepath, Integer parentorganizationId, @NotNull String name) {
+    public Direction(Long id, Long organizationId, Long level, String type, String treepath, Long parentorganizationId, @NotNull String name) {
         this.id = id;
         this.organizationId = organizationId;
         this.level = level;
@@ -111,11 +111,11 @@ public class Direction {
         this.treepath = treepath;
     }
 
-    public Integer getParentorganizationId() {
+    public Long getParentorganizationId() {
         return parentorganizationId;
     }
 
-    public void setParentorganizationId(Integer parentorganizationId) {
+    public void setParentorganizationId(Long parentorganizationId) {
         this.parentorganizationId = parentorganizationId;
     }
